Calculate dynamics alignment in a sane, configurable way.
Until now, DynamicText alignment was a mess:

The correct way to align a grob is to assign an appropriate
callback to the grob's offset property; LilyPond should
evaluate that callback and use its return value as grob's
offset when calculating its position.  Default settings
should be specified in define-grobs.scm; users should be
able to override them by assigning another procedure (or
a number) to that property.

DynamicTexts, however, didn't conform to this standard: one
part of alignment was in define-grobs.scm, and another one
on the C++ level: Dynamic_engraver and New_dynamic_engraver
applied set_center_parent on all DynamicTexts they created.
This way, regardless of what X-offset property was set to,
half a NoteHead width was always added to it.

This commit deletes calls to set_center_parent; DynamicText's
alignment is now defined in one place, with no hard-coded
C++ stuff.

This commit shouldn't change anything in Lily behaviour.

See also next commit, which does the same for Fingerings.