Use align-grob in define-grobs.scm: AmbitusLine
TODO: Apparently there is something wrong with AmbitusLine,
because a warning "AmbitusLine has empty extent and non-empty
stencil" pops out.  Obviously, AmbitusLine's stencil isn't
empty, because there's a line to be printed.  It seems that
LilyPond fails to set appropriate extent, which should be
automatically calculated from the stencil.

See also other commits in this branch.