use X-parents to align MMRNumbers, MMRTexts and PercentRepeatCounters
MultiMeasureRestNumber, MultiMeasureRestText and PercentRepeatCounter
are spanners (the grobs they are attached to are spanners as well, and
can be broken over barlines, so these three also must be spanners).

Until now their horizontal alignment was calculated quite unusually:
they were centered on Y-parents (MultiMeasureRest and PercentRepeat,
respectively). What's unusual about this is that it would make sense
to use X-parents for horizontal alignment - set MMRNumber's X-parent
to MMR and simply align horizontally on X-parent.  That's how the
alignment is specified now.

This makes x_centered_on_y_parent obsolete - it will be removed.

This commit shouldn't change anything in Lily behaviour.