Make sure slurs actually avoid stafflines.
Until now, slurs were pretending that they were avoiding stafflines,
but the function was broken: it only moved slurs in one direction
(sometimes making the collision signifincantly worse).

This commit makes sure that the correction is applied in appropriate
direction and that slur thickness is taken into account. Additionally,
the amount of correction is now user-settable.