Don't put RehearsalMarks in a NonMusicalPaperColumn (fix 3875)
Until now, RehearsalMark was kept in a column separate from musical notation
(a NonMusicalPaperColumn) - this behaviour was apparently introduced in commit
e03b2923c.

The result was that inserting a RehearsalMark inside a measure, like this
{ r4 r r \mark I r }
resulted in extra horizontal space being added.  This patch makes RehearsalMarks
a part of ordinary (musical) PaperColumn, so that no extra space is added.