Issue 3299: better alignment with PaperColumn parent
PaperColumns are the objects that define where a musical moment
"happens" on the staff horizontally.  However, the noteheads aren't
center-aligned on PaperColumns - according to notation rules they
are left-aligned.  Because of that an object centered on a notehead
is not actually centered on the PaperColumn, which results in a
discrepancy between objects attached to NoteHeads and PaperColumns.

This commit fixes that by introducing a placeholder extent which is
used for aligning such grobs.